keyboards: fr: fix keyboard layout
A number of keys would not give the expected result, due to the lack of
a `text` of `action` property. Additionally, the layout wouldn't load
because of an old file format (it seems `bounds` are not supported
anymore).

This patch makes sure this layout can be loaded and used by squeekboard
v0.1.9.